Ignore nested type with invalid class file
Prior to this commit, an attempt to find nested test classes using
ReflectionUtils.findNestedClasses() resulted in a NoClassDefFoundError
if a nested class or nested interface had an invalid class file (which
could happen due to a bug in a compiler such as one reported for the
Groovy compiler).

This commit addresses this issue by catching a NoClassDefFoundError
thrown by an invocation of Class#getDeclaredClasses() and logging it
instead of allowing it to propagate. Such nested classes and their
siblings are now effectively ignored.

Issue: #1436